tapcfg direct access to the file descriptor of the interface
Summary:
tapcfg provides read/write and wait functions but this is incompatible with our usage of libevent.

We would like to get a direct access to the file descriptor of the tun interface.

Test Plan:
Should work correctly on all Unix
The Windows version has to be tested

Reviewers: Chipot

Reviewed By: Chipot

CC: Zerk

Maniphest Tasks: T1

Differential Revision: http://trac.medu.se/D1